museboard-ai-service/requirements.txt
git+https://github.com/stanfordnlp/dspy.git
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
openai